NeuralBrain-AI/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy>=2.0.46
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
PyJWT==2.8.0
urllib3>=2.0.0
# scikit-learn==1.3.2
# numpy==1.24.3
# pandas==2.0.3
python-dateutil==2.8.2

psycopg[binary]
